requirements.txt
# Data handling and analysis
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Data fetching
yfinance==0.2.36
fredapi==0.6.1
requests==2.31.0

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Model interpretability
shap==0.43.0

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0 